backend/requirements.txt
fastapi[standard]
sqlalchemy
pydantic-settings
redis
psycopg
psycopg2-binary
alembic